sisu_bot/requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
rapidfuzz==3.6.2
SQLAlchemy>=2.0.0
aiosqlite>=0.19.0
python-json-logger>=2.0.0
redis>=5.0.0
aioredis>=2.0.0
pydantic>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.0.0
flake8>=6.0.0
mypy>=1.0.0
